+++ demo/src/main/java/net/vrallev/android/task/demo/MainActivity.java
@@ -1,6 +1,7 @@
 package net.vrallev.android.task.demo;
 
 import android.app.Activity;
+import android.app.AlertDialog;
 import android.app.ProgressDialog;
 import android.content.DialogInterface;
 import android.content.Intent;
@@ -9,14 +10,20 @@ import android.support.annotation.NonNull;
 import android.view.View;
 import android.widget.Toast;
 
+import net.vrallev.android.task.Task;
 import net.vrallev.android.task.TaskExecutor;
 import net.vrallev.android.task.TaskResult;
+
+import java.util.Iterator;
+import java.util.List;
 
 /**
  * @author rwondratschek
  */
 @SuppressWarnings("UnusedDeclaration")
 public class MainActivity extends Activity {
+
+    private static final String ANNOTATION_ID = "annotationId";
 
     private static final String TASK_ID_KEY = "TASK_ID_KEY";
 
@@ -72,6 +79,14 @@ public class MainActivity extends Activity {
                 testIntegerTask();
                 break;
 
+            case R.id.button_annotation_with_id:
+                testAnnotationWithId();
+                break;
+
+            case R.id.button_get_all_tasks:
+                testGetAllTasks();
+                break;
+
             case R.id.button_shutdown_executor:
                 testShutdownExecutor();
                 break;
@@ -90,8 +105,13 @@ public class MainActivity extends Activity {
     }
 
     @TaskResult
-    public void onResult(Integer integer) {
+    public void onResult(Integer integer, IntegerTask task) {
         Toast.makeText(this, "Result " + integer, Toast.LENGTH_SHORT).show();
+    }
+
+    @TaskResult(id = ANNOTATION_ID)
+    public void onResultWithId(Integer integer, Task<?> task) {
+        Toast.makeText(this, "Result with ID " + integer + ", finished = " + task.isFinished(), Toast.LENGTH_SHORT).show();
     }
 
     private void testDefaultActivity() {
@@ -107,6 +127,29 @@ public class MainActivity extends Activity {
         TaskExecutor.getInstance().execute(new IntegerTask(), this);
     }
 
+    private void testAnnotationWithId() {
+        TaskExecutor.getInstance().execute(new IntegerTask(), this, ANNOTATION_ID);
+    }
+
+    private void testGetAllTasks() {
+        List<IntegerTask> allTasks = TaskExecutor.getInstance().getAllTasks(IntegerTask.class);
+        Iterator<IntegerTask> iterator = allTasks.iterator();
+
+        StringBuilder stringBuilder = new StringBuilder();
+        while (iterator.hasNext()) {
+            stringBuilder.append(iterator.next());
+            if (iterator.hasNext()) {
+                stringBuilder.append("\n\n");
+            }
+        }
+
+        new AlertDialog.Builder(this)
+                .setTitle("Tasks")
+                .setMessage(stringBuilder)
+                .setPositiveButton(android.R.string.ok, null)
+                .show();
+    }
+
     private void testShutdownExecutor() {
         TaskExecutor.getInstance().shutdown();
     }
@@ -119,7 +162,7 @@ public class MainActivity extends Activity {
             @Override
             public void onClick(DialogInterface dialog, int which) {
                 if (mTaskId != -1) {
-                    SimpleTask task = TaskExecutor.getInstance().getTask(mTaskId);
+                    SimpleTask task = (SimpleTask) TaskExecutor.getInstance().getTask(mTaskId);
                     if (task != null) {
                         task.cancel();
                     }