+++ library/src/main/java/net/vrallev/android/task/TaskCacheFragment.java
@@ -5,13 +5,10 @@ import android.app.Activity;
 import android.app.Fragment;
 import android.app.FragmentManager;
 import android.os.Bundle;
-import android.util.Pair;
 
-import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 
@@ -71,18 +68,15 @@ public final class TaskCacheFragment extends Fragment implements TaskCacheFragme
         mCanSaveInstanceState = true;
 
         List<TaskPendingResult> list = get(PENDING_RESULT_KEY);
-        if (list != null) {
-            Iterator<TaskPendingResult> iterator = list.iterator();
-            TargetMethodFinder targetMethodFinder = new TargetMethodFinder(TaskResult.class);
-
-            while (iterator.hasNext()) {
-                TaskPendingResult pendingResult = iterator.next();
-                iterator.remove();
-
-                Pair<Method, Object> target = targetMethodFinder.getMethod(this, pendingResult.getResultType());
-                targetMethodFinder.invoke(target, pendingResult.getResult());
-            }
+        if (list != null && !list.isEmpty()) {
+            TaskCacheFragmentInterface.Helper.postPendingResults(list, this);
         }
+    }
+
+    @Override
+    public void onStop() {
+        mCanSaveInstanceState = false;
+        super.onStop();
     }
 
     @Override