+++ library/src/main/java/net/vrallev/android/task/TaskPendingResult.java
@@ -7,10 +7,14 @@ public class TaskPendingResult {
 
     private final Class<?> mResultType;
     private final Object mResult;
+    private final Task<?> mTask;
+    private final TaskExecutor mTaskExecutor;
 
-    public TaskPendingResult(Class<?> resultType, Object result) {
+    public TaskPendingResult(Class<?> resultType, Object result, Task<?> task, TaskExecutor taskExecutor) {
         mResultType = resultType;
         mResult = result;
+        mTask = task;
+        mTaskExecutor = taskExecutor;
     }
 
     public Class<?> getResultType() {
@@ -19,5 +23,13 @@ public class TaskPendingResult {
 
     public Object getResult() {
         return mResult;
+    }
+
+    public Task<?> getTask() {
+        return mTask;
+    }
+
+    public TaskExecutor getTaskExecutor() {
+        return mTaskExecutor;
     }
 }