+++ library/src/main/java/net/vrallev/android/task/TaskResult.java
@@ -8,4 +8,5 @@ import java.lang.annotation.RetentionPolicy;
  */
 @Retention(RetentionPolicy.RUNTIME)
 public @interface TaskResult {
+    String id() default "";
 }