+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 1.0.3 (2015-03-12)
+
+Bugfixes:
+
+  - add Annotation ID for concrete callback method
+  - add Task reference in callback method
+  - add option to resolve all pending tasks
+  - bug fixes
+
 ## 1.0.2 (2015-03-08)
 
 Bugfixes: